Modify how font size is applied to fourth element
I had applied the 24px font size to the fourth element by adding the id
selector "fourth" to the "font-size: 24px;" rule in the CSS file.

In the solution, instead of the "font-size: 24px;" rule having two
selectors (id selector "fourth" and class selector ".bigger-font"),
the fourth element has two attributes (both an "id" and a "class")
in the HTML file.

While my original solution did achieve the same outcome, I think having
the class "bigger-font" attribute on the fourth element
in the HTML file is clearer than adding the id selector "fourth" to the
"font-size: 24px" rule in the CSS file. So I mimicked the solution.